keelson-record/requirements.txt
eclipse-zenoh==0.10.0-rc
mcap==1.1.1
mcap-protobuf-support==0.4.0
parse==1.19.1